Anti Theft Alarm is basically an Android application.
An Android application that uses proximity sensor and safeguard mobile
from theft.
On opening the application,the user is asked to enter a password.Then
the proximity sensor starts to get values.If the mobile is placed in the
Users pocket no harm;else the alarm starts to ring until the user enters
the same password entered by him during login.